requirements.txt
Django==5.2
python-dotenv==1.1.0
PyJWT==2.10.1
django-allauth==65.7.0
psycopg2==2.9.10
djangorestframework==3.16.0
django-multitenant==4.1.1
cryptography==44.0.3
# pip install  --no-cache-dir django_multitenant
# pip install cryptography
# pip install pillow
# pip install django-cors-headers
# pip install dj-rest-auth